Fixed region not being set correctly and legal description not being shown
