+++ source/frontend/src/hooks/useBcAssessmentLayer.ts
@@ -1,9 +1,8 @@
 import axios, { AxiosError, AxiosResponse } from 'axios';
 import { IApiError } from 'interfaces/IApiError';
-import { getMockAddresses, getMockLegalDescriptions } from 'mocks/bcAssessmentMock';
+import { mockBcAssessmentSummary } from 'mocks/bcAssessmentMock';
 import { useCallback, useEffect } from 'react';
 
-import { getMockDescription, getMockSales, getMockValues } from './../mocks/bcAssessmentMock';
 import { pidParser } from './../utils/propertyUtils';
 import { useWfsLayer } from './pims-api';
 import { IResponseWrapper, useApiRequestWrapper } from './pims-api/useApiRequestWrapper';
@@ -236,14 +235,6 @@ export const useBcAssessmentLayer = (
   };
 };
 
-export const mockBcAssessmentSummary: IBcAssessmentSummary = {
-  LEGAL_DESCRIPTION: getMockLegalDescriptions()?.features[0].properties ?? {},
-  ADDRESSES: getMockAddresses()?.features.map(f => f.properties ?? {}) ?? [],
-  VALUES: getMockValues()?.features.map(f => f.properties ?? {}) ?? [],
-  CHARGES: [],
-  FOLIO_DESCRIPTION: (getMockDescription()?.features[0].properties as any) ?? {},
-  SALES: getMockSales()?.features.map(f => f.properties ?? {}) ?? [],
-};
 export interface IBcAssessmentSummary {
   FOLIO_DESCRIPTION: Partial<{
     BCA_FD_SYSID: number;
